\name{prune.size.testsample}
\alias{prune.size.testsample}
\title{Prunes and Selects Best-Sized Multivariate Survival Tree}
\description{Prunes the initial tree and selects the best-sized multivariate survival tree}
\usage{prune.size.testsample(tree)}
\arguments{
  \item{tree}{The initial tree}
}
\value{
  \item{result}{Information on trees pruned}
  \item{btree}{The best-sized tree structures}
}
\author{Xiaogang Su, Peter Calhoun, Juanjuan Fan}
\seealso{\code{\link{MST}}}
\keyword{Trees}
\keyword{Survival}
\keyword{Multivariate}
\keyword{Correlated}
